Add koios node to mainnet
### Team:
Koios
My discord handle is CryingEaglez#6584

### Use case:
We will use ceramic for our user (student/lecturer) profiles. We are planning to make use of the tileDocument streamtype for storing custom user data (stats for instance) so they can move between mobile and pc platforms and also exchange info with (and find) other users.

### Overview:
We are running our ceramic node on a locally hosted server. IPFS is ran out of process to have more control over the options, resource allocation, maintenance, debugging and observability.

*Multiaddress persistence:*
The multiaddress is stored locally and physically backed up.

*Ceramic State Store persistence:*
The ceramic state store and IPFS repo are stored locally and are hourly backed up to a NAS (with zfs)

*IPFS Repo persistence:*
The ceramic state store and IPFS repo are stored locally and are hourly backed up to a NAS (with zfs)

*Static IP:*
217.102.168.211